# pkg/api/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.0 DO NOT EDIT.
package openapi

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// NewPet defines model for NewPet.
type AllocateAll struct {
	// Name Name of the pet
	ReservationIds *[]int32 `json:"reservationIds"`

	// Tag Type of the pet
	UserId *int32 `json:"userId,omitempty"`
}

// Pet defines model for Pet.
type Pet struct {
	// Id Unique id of the pet
	Id int64 `json:"id"`

	// Name Name of the pet
	Name string `json:"name"`

	// Tag Type of the pet
	Tag *string `json:"tag,omitempty"`
}

// FindPetsParams defines parameters for FindPets.
type FindPetsParams struct {
	// Tags tags to filter by
	Tags *[]string `form:"tags,omitempty" json:"tags,omitempty"`

	// Limit maximum number of results to return
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`
}

type AllocateAllParams struct {
	// Tags tags to filter by
	ReservationIDs *[]int32 `form:"reservationIDs,omitempty" json:"reservationIDs,omitempty"`

	// Limit maximum number of results to return
	UserID *int32 `form:"userID,omitempty" json:"userID,omitempty"`
}

// AddPetJSONRequestBody defines body for AddPet for application/json ContentType.
type AddPetJSONRequestBody = AllocateAll

type AutoAllocateRequestBody struct {
	ReservationID int32 `json:"reservationID"`
	AgentID         int32   `json:"agentID"`
	IsNotify         bool   `json:"isNotify"`
}

type AllocateAllRequestBody struct {
	ReservationIDs []int32 `json:"reservationIDs"`
	UserID         int32   `json:"userID"`
}